Refactor store to use modules for better organization, update relevant components to use new store structure
