+++ src/App.vue
@@ -22,7 +22,7 @@ export default {
     };
   },
   methods: {
-    ...mapActions(["fetchAuthUser"]),
+    ...mapActions("auth", ["fetchAuthUser"]),
     onPageReady() {
       this.showPage = true;
       NProgress.done();
