+++ src/components/PostList.vue
@@ -34,7 +34,7 @@
           </div>
         </div>
         <a
-          v-if="post.userId === $store.state.authId"
+          v-if="post.userId === $store.state.auth.authId"
           @click.prevent="toggleEditMode(post.id)"
           href="#"
           style="margin-left: auto;"
@@ -73,13 +73,13 @@ export default {
   },
   computed: {
     users() {
-      return this.$store.state.users;
+      return this.$store.state.users.items;
     }
   },
   methods: {
-    ...mapActions(["updatePost"]),
+    ...mapActions("posts", ["updatePost"]),
     userById(userId) {
-      return this.$store.getters.user(userId);
+      return this.$store.getters["users/user"](userId);
     },
     toggleEditMode(id) {
       this.editing = id === this.editing ? null : id;
