+++ src/components/UserProfileCardEditor.vue
@@ -99,7 +99,7 @@ export default {
   },
   methods: {
     save() {
-      this.$store.dispatch("updateUser", { ...this.activeUser });
+      this.$store.dispatch("users/updateUser", { ...this.activeUser });
       this.$router.push({ name: "Profile" });
     },
     cancel() {
