+++ src/helpers/index.js
@@ -26,3 +26,22 @@ export const docToResource = doc => {
     ...doc.data()
   };
 };
+
+export const makeAppendChildToParentMutation = ({ parent, child }) => {
+  return (state, { childId, parentId }) => {
+    const resource = findById(state.items, parentId);
+
+    if (!resource) {
+      console.warn(
+        `Appending ${child} ${childId} to ${parent} ${parentId} failed because the parent didn't exist`
+      );
+      return;
+    }
+
+    resource[child] = resource[child] || [];
+
+    if (!resource[child].includes(childId)) {
+      resource[child].push(childId);
+    }
+  };
+};
