+++ src/pages/Category.vue
@@ -24,13 +24,14 @@ export default {
   mixins: [asyncDataStatus],
   computed: {
     category() {
-      return findById(this.$store.state.categories, this.id) || {};
+      return findById(this.$store.state.categories.items, this.id) || {};
     }
   },
   methods: {
-    ...mapActions(["fetchCategory", "fetchForums"]),
+    ...mapActions("categories", ["fetchCategory"]),
+    ...mapActions("forums", ["fetchForums"]),
     getCategoryForums(category) {
-      return this.$store.state.forums.filter(
+      return this.$store.state.forums.items.filter(
         forum => forum.categoryId === category.id
       );
     }
