+++ src/pages/Forum.vue
@@ -38,7 +38,7 @@ export default {
   mixins: [asyncDataStatus],
   computed: {
     forum() {
-      return findById(this.$store.state.forums, this.id);
+      return findById(this.$store.state.forums.items, this.id);
     },
     threads() {
       if (!this.forum) {
@@ -47,13 +47,15 @@ export default {
 
       return this.forum.threads
         ? this.forum.threads.map(threadId =>
-            this.$store.getters.thread(threadId)
+            this.$store.getters["threads/thread"](threadId)
           )
         : [];
     }
   },
   methods: {
-    ...mapActions(["fetchForum", "fetchThreads", "fetchUsers"])
+    ...mapActions("forums", ["fetchForum"]),
+    ...mapActions("threads", ["fetchThreads"]),
+    ...mapActions("users", ["fetchUsers"])
   },
   async created() {
     const forum = await this.fetchForum({ id: this.id });
