+++ src/pages/Home.vue
@@ -17,11 +17,12 @@ export default {
   mixins: [asyncDataStatus],
   computed: {
     categories() {
-      return this.$store.state.categories;
+      return this.$store.state.categories.items;
     }
   },
   methods: {
-    ...mapActions(["fetchAllCategories", "fetchForums"])
+    ...mapActions("categories", ["fetchAllCategories"]),
+    ...mapActions("forums", ["fetchForums"])
   },
   async created() {
     const categories = await this.fetchAllCategories();
