+++ src/pages/Profile.vue
@@ -41,10 +41,10 @@ export default {
   },
   mixins: [asyncDataStatus],
   computed: {
-    ...mapGetters({ user: "authUser" })
+    ...mapGetters("auth", { user: "authUser" })
   },
   async created() {
-    await this.$store.dispatch("fetchAuthUserPosts");
+    await this.$store.dispatch("auth/fetchAuthUserPosts");
     this.asyncDataStatus_fetched();
   }
 };
