+++ src/pages/ThreadCreate.vue
@@ -37,11 +37,12 @@ export default {
   },
   computed: {
     forum() {
-      return findById(this.$store.state.forums, this.id);
+      return findById(this.$store.state.forums.items, this.id);
     }
   },
   methods: {
-    ...mapActions(["createThread", "fetchForum"]),
+    ...mapActions("threads", ["createThread"]),
+    ...mapActions("forums", ["fetchForum"]),
     async save({ title, text }) {
       const thread = await this.createThread({
         forumId: this.forum.id,
