+++ src/pages/ThreadShow.vue
@@ -58,28 +58,24 @@ export default {
   },
   mixins: [asyncDataStatus],
   computed: {
-    ...mapGetters(["authUser"]),
+    ...mapGetters("auth", ["authUser"]),
     thread() {
-      return this.$store.getters.thread(this.id);
+      return this.$store.getters["threads/thread"](this.id);
     },
     threadPosts() {
       return this.posts.filter(post => post.threadId === this.id);
     },
     threads() {
-      return this.$store.state.threads;
+      return this.$store.state.threads.items;
     },
     posts() {
-      return this.$store.state.posts;
+      return this.$store.state.posts.items;
     }
   },
   methods: {
-    ...mapActions([
-      "createPost",
-      "fetchThread",
-      "fetchUser",
-      "fetchPosts",
-      "fetchUsers"
-    ]),
+    ...mapActions("posts", ["createPost", "fetchPosts"]),
+    ...mapActions("threads", ["fetchThread"]),
+    ...mapActions("users", ["fetchUsers"]),
     addPost(eventData) {
       const post = {
         ...eventData.post,
