+++ src/router/index.js
@@ -67,8 +67,8 @@ const routes = [
     component: ThreadShow,
     props: true,
     async beforeEnter(to, from, next) {
-      await store.dispatch("fetchThread", { id: to.params.id });
-      const threadExists = findById(store.state.threads, to.params.id);
+      await store.dispatch("threads/fetchThread", { id: to.params.id });
+      const threadExists = findById(store.state.threads.items, to.params.id);
 
       if (!threadExists) {
         return next({
@@ -111,7 +111,7 @@ const routes = [
     path: "/logout",
     name: "LogOut",
     async beforeEnter(to, from) {
-      await store.dispatch("logOut");
+      await store.dispatch("auth/logOut");
 
       return { name: "Home" };
     }
@@ -142,17 +142,17 @@ const router = createRouter({
 });
 
 router.beforeEach(async to => {
-  await store.dispatch("initAuthentication");
+  await store.dispatch("auth/initAuthentication");
   store.dispatch("unsubscribeAllSnapshots");
 
-  if (to.meta.requiresAuth && !store.state.authId) {
+  if (to.meta.requiresAuth && !store.state.auth.authId) {
     return {
       name: "LogIn",
       query: { redirectTo: to.path }
     };
   }
 
-  if (to.meta.requiresGuest && store.state.authId) {
+  if (to.meta.requiresGuest && store.state.auth.authId) {
     return { name: "Home" };
   }
 });
