+++ src/store/mutations.js
@@ -1,33 +1,8 @@
-import { findById, upsert, docToResource } from "@/helpers";
+import { upsert, docToResource } from "@/helpers";
 
 export default {
   setItem(state, { resource, item }) {
-    upsert(state[resource], docToResource(item));
-  },
-  setAuthId(state, id) {
-    state.authId = id;
-  },
-  appendPostToThread: makeAppendChildToParentMutation({
-    parent: "threads",
-    child: "posts"
-  }),
-  appendThreadToForum: makeAppendChildToParentMutation({
-    parent: "forums",
-    child: "threads"
-  }),
-  appendThreadToUser: makeAppendChildToParentMutation({
-    parent: "users",
-    child: "threads"
-  }),
-  appendContributorToThread: makeAppendChildToParentMutation({
-    parent: "threads",
-    child: "contributors"
-  }),
-  setAuthUserUnsubscribe(state, unsubscribe) {
-    state.authUserUnsubscribe = unsubscribe;
-  },
-  setAuthObserverUnsubscribe(state, unsubscribe) {
-    state.authObserverUnsubscribe = unsubscribe;
+    upsert(state[resource].items, docToResource(item));
   },
   appendUnsubscribe(state, { unsubscribe }) {
     state.unsubscribes.push(unsubscribe);
@@ -36,22 +11,3 @@ export default {
     state.unsubscribes = [];
   }
 };
-
-function makeAppendChildToParentMutation({ parent, child }) {
-  return (state, { childId, parentId }) => {
-    const resource = findById(state[parent], parentId);
-
-    if (!resource) {
-      console.warn(
-        `Appending ${child} ${childId} to ${parent} ${parentId} failed because the parent didn't exist`
-      );
-      return;
-    }
-
-    resource[child] = resource[child] || [];
-
-    if (!resource[child].includes(childId)) {
-      resource[child].push(childId);
-    }
-  };
-}
